updated code to work with new json parsing and template context

+++ app/src/main/java/com/scouting_app_2026/UIElements/SliderElement.java
@@ -24,7 +24,7 @@ public class SliderElement extends UIElement {
 
     @Override
     public String getValue() {
-        return Float.toString(slider.getValue());
+        return Integer.toString((int)slider.getValue());
     }
 
     public void setLabels(String[] labels) {
